# Bug-Tracker/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotSessionStatus = typeof BotSessionStatus[keyof typeof BotSessionStatus];


export const BotSessionStatus = {
  spawning: 'spawning',
  running: 'running',
  stopped: 'stopped',
  failed: 'failed',
} as const;

export interface BotSession {
  id: string;
  host: string;
  port: number;
  botCount: number;
  delayMs: number;
  status: BotSessionStatus;
  connectedCount: number;
  failedCount: number;
  errorMessage: string;
  createdAt: string;
}

export interface LaunchInput {
  /** @minLength 1 */
  host: string;
  /** @nullable */
  port?: number | null;
  /**
     * @minimum 1
     * @maximum 500
     */
  botCount: number;
  /** @minimum 0 */
  delayMs: number;
}

export interface StopResult {
  success: boolean;
  message: string;
}

export interface BotStats {
  totalSessions: number;
  activeSessions: number;
  totalBotsLaunched: number;
  totalBotsConnected: number;
}

export type UpdateSessionBody = {
  host?: string;
  port?: number;
  botCount?: number;
  delayMs?: number;
};