# Bug-Tracker/lib/api-zod/src/generated/types/botSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotSessionStatus } from './botSessionStatus';

export interface BotSession {
  id: string;
  host: string;
  port: number;
  botCount: number;
  delayMs: number;
  status: BotSessionStatus;
  connectedCount: number;
  failedCount: number;
  errorMessage: string;
  createdAt: string;
}