# Bug-Tracker/lib/api-zod/src/generated/types/botSessionStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type BotSessionStatus = typeof BotSessionStatus[keyof typeof BotSessionStatus];


export const BotSessionStatus = {
  spawning: 'spawning',
  running: 'running',
  stopped: 'stopped',
  failed: 'failed',
} as const;