# Bug-Tracker/lib/api-zod/src/generated/types/botStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotStats {
  totalSessions: number;
  activeSessions: number;
  totalBotsLaunched: number;
  totalBotsConnected: number;
}