# Bug-Tracker/lib/api-zod/src/generated/types/launchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LaunchInput {
  /** @minLength 1 */
  host: string;
  /** @nullable */
  port?: number | null;
  /**
     * @minimum 1
     * @maximum 500
     */
  botCount: number;
  /** @minimum 0 */
  delayMs: number;
}